Migrated parsing logic to SemanticFieldMapper
